Final Project
Project executed in the week of September 11th, with material provided by RocketSeat in the NLW Esports project.
In this project I carried out the development of a Basic Website with the intention of reviewing the knowledge I obtained so far in HTML5/CSS3, in addition to being able to include another amazing project in my Portfolio.

The Project consisted of creating a web page with interactive links for direct access to some personal information, such as Favorite Games, Streamers I watch the most, and a brief link to my social networks.